requirements.txt
tqdm==4.47.0
pytorch_warmup==0.0.4
thop==0.0.31.post2005241907
graphviz==0.14.1
spacy==2.3.1
pytorch_pretrained_bert==0.6.2
ujson==3.0.0
h5py==3.1.0
opencv_python==4.3.0.36
pandas==1.0.5
torch==1.4.0
scikit_video==1.1.11
matplotlib==3.2.2
numpy==1.19.5
easydict==1.9
fire==0.3.1
colorlog==4.0.2
prefetch_generator==1.0.1
torchvision==0.5.0
allennlp==2.5.0
nltk==3.6.2
Pillow==8.2.0
tensorboardX==2.2
